requirements.txt
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
Deprecated==1.2.14
googleapis-common-protos==1.59.1
idna==3.4
importlib-metadata==6.8.0
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp-proto-common==1.19.0
opentelemetry-exporter-otlp-proto-http==1.19.0
opentelemetry-instrumentation==0.40b0
opentelemetry-instrumentation-system-metrics==0.40b0
opentelemetry-proto==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-semantic-conventions==0.40b0
protobuf==4.23.4
psutil==5.9.5
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
zipp==3.16.2
